Tweak to default macro name
Apologies for making such a tiny change, but this has always bugged me.

It took me embarrasingly long to figure out that "E-Stop" stands for
"Emergency-Stop" (my mind jumped straight to Extruder-Stop - which made
no sense, I know!).

I figure, why not be explicit?

There is plenty of space for the full word, since the sidebar is hardcoded to
300px as far as I can tell, and in the klipper tab we have a button for "Assisted Bed Leveling", which is significantly longer.